third_party/blink/web_tests/wpt_internal/declarative-performance-observer/capture-early-failures.https.sub.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Declarative Performance Observer: capture-early-failures</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/reporting/resources/report-helper.js"></script>
<body>
<h1>Declarative Performance Observer: Capture Early Failures</h1>
<script>
promise_test(async t => {
  const reportID = crypto.randomUUID();
  const base_url = `${location.protocol}//${location.host}`;
  const endpoint = `${base_url}/reporting/resources/report.py`;

  const readyPromise = new Promise(resolve => {
    const listener = e => {
      if (e.data === "ready") {
        window.removeEventListener("message", listener);
        resolve();
      }
    };
    window.addEventListener("message", listener);
  });

  // 1. Opt-in Window
  const optInURL = `${base_url}/wpt_internal/declarative-performance-observer/resources/capture-early-failures-opt-in.https.sub.html?reportID=${reportID}`;
  const popup = window.open(optInURL);
  assert_not_equals(popup, null, "Opt-in popup should be opened");
  t.add_cleanup(() => { popup.close(); });
  await readyPromise;

  // 2. Trigger a sleeping navigation
  const earlyFailureURL = `${base_url}/wpt_internal/declarative-performance-observer/resources/early-failure.py`;
  popup.location.href = earlyFailureURL;

  // Wait 100ms to ensure the browser process has initiated the navigation.
  await new Promise(resolve => t.step_timeout(resolve, 100));

  const collectReady = new Promise(resolve => {
    const listener = e => {
      if (e.data === "collect_committed") {
        window.removeEventListener("message", listener);
        resolve();
      }
    };
    window.addEventListener("message", listener);
  });

  // 3. Abort the sleeping navigation by navigating to the collector.
  // The C++ navigation pipeline guarantees that the abortion of the previous
  // navigation is fully processed and written to the database BEFORE the
  // new collector page commits and reads it.
  const collectURL = `${base_url}/wpt_internal/declarative-performance-observer/resources/capture-early-failures-collect.https.sub.html?reportID=${reportID}`;
  popup.location.href = collectURL;
  await collectReady;

  // 4. Poll the reporting endpoint, blocking until at least 1 report is received.
  const reports = await pollReports(endpoint, reportID, 1);

  assert_greater_than_equal(reports.length, 1, "Should receive early failure report");
  const report = reports[0];
  assert_equals(report.type, "performance-observer");
  const entries = report.body.entries;
  assert_true(!!entries && entries.length >= 1, "Report should have navigation failure entries");
  const entry = entries[0];
  assert_equals(entry.entryType, "navigation");
}, "Early navigation failure reports are successfully buffered and flushed to the reporting endpoint");
</script>
</body>
